Source code/requirements.txt
pandas
numpy
scikit-learn
matplotlib
python-docx
python-pptx
Pillow
datasets
paddleocr
paddlepaddle
pyarrow
PyYAML
